Rename main page components to index
