+++ frontend/src/App.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Home from './pages/Home/Home';
-import DatasetExplorer from './pages/DatasetExplorer/DatasetExplorer';
-import AILab from './pages/AILab/AILab';
-import Solutions from './pages/Solutions/Solutions'; // ✅ ¡Así!
+import Home from './pages/Home';
+import DatasetExplorer from './pages/DatasetExplorer';
+import AILab from './pages/AILab';
+import Solutions from './pages/Solutions';
 
 export default function App() {
   return (
