+++ frontend/src/pages/AILab/README.md
@@ -1,0 +1,3 @@
+# AILab Page
+
+This directory houses the React components and logic for the **AILab** page.
