+++ frontend/src/pages/AILab/index.jsx
