+++ frontend/src/pages/DatasetExplorer/README.md
@@ -1,0 +1,3 @@
+# DatasetExplorer Page
+
+This directory houses the React components and logic for the **DatasetExplorer** page.
