+++ frontend/src/pages/Home/README.md
@@ -1,0 +1,3 @@
+# Home Page
+
+This directory houses the React components and logic for the **Home** page.
