+++ frontend/src/pages/Solutions/README.md
@@ -1,0 +1,3 @@
+# Solutions Page
+
+This directory houses the React components and logic for the **Solutions** page.
