+++ frontend/src/pages/Solutions/index.jsx
